W12hw1 webpack-dev-server, react-hot-loader
