+++ homeworks/week12/hw1/.babelrc
@@ -1,3 +1,4 @@
 {
-  "presets": ["env", "react"]
+  "presets": ["env", "react"],
+  "plugins": ["react-hot-loader/babel"]
 }
