+++ homeworks/week12/hw1/webpack.config.js
@@ -10,8 +10,9 @@ module.exports = {
       loader:"babel-loader" }
     ]
   },
-  output:{
+  output: {
     filename: 'bundle.js',
-    path: path.resolve(__dirname, 'dist')
+    path: path.resolve(__dirname, 'dist'),
+    publicPath: '/dist'
   }
 };
